+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+## [0.3.0] - 2019-12-18
+* tweaked bevelled bars
+* **BREAKING CHANGE**: changed `thresholds` to `levels`
+* added **sector signal indicator** (`SignalStrengthIndicator.sector`)
+
 ## [0.2.0] - 2019-12-02
 * added `bevelled` option to bar signal indicator
 
 ## [0.1.0] - 2019-11-29
 * first release
-* **bar signal indicator**
+* **bar signal indicator** (`SignalStrengthIndicator.bars`)